move types into decoder and sync waf disabled api
